requirements.txt
scipy==1.4.1
matplotlib==3.3.0
numpy==1.18.5
py-ecg-detectors==1.0.2
pandas==1.1.0
mysql-connector-python==8.0.26
xgboost==1.5.0